requirements.txt
accelerate==0.20.3
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
appdirs==1.4.4
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blobfile==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
docker-pycreds==0.4.0
evals==0.1.1
exceptiongroup==1.1.1
fastapi==0.96.0
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
fonttools==4.39.4
frozenlist==1.3.3
fschat==0.2.11
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
gradio==3.23.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langdetect==1.0.9
linkify-it-py==2.0.2
lit==16.0.5.post0
lxml==4.9.2
lz4==4.3.2
markdown-it-py==2.2.0
markdown2==2.4.8
MarkupSafe==2.1.3
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mock==5.0.2
mpmath==1.3.0
multidict==6.0.4
mypy==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nh3==0.2.13
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.27.7
orjson==3.9.0
oscrypto==1.3.0
packaging==23.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.5.0
pip==23.0.1
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
pyarrow==10.0.1
pycparser==2.21
pycryptodomex==3.18.0
pydantic==1.10.8
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzstd==0.15.7
regex==2023.6.3
requests==2.31.0
rich==13.4.1
sacrebleu==2.3.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.25.1
setproctitle==1.3.2
setuptools==67.8.0
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowflake-connector-python==3.0.4
sortedcontainers==2.4.0
starlette==0.27.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tqdm==4.65.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.6.3
typing-inspect==0.9.0
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==1.26.16
uvicorn==0.22.0
wandb==0.15.4
wavedrom==2.0.3.post3
wcwidth==0.2.6
websockets==11.0.3
wheel==0.38.4
xformers==0.0.20
yarl==1.9.2
zipp==3.15.0
